Move to ruff (issue #1014)
