+++ .github/workflows/ibllib_ci.yml
@@ -35,12 +35,12 @@ jobs:
     - name: Install deps
       run: |
         python -m pip install --upgrade pip
-        python -m pip install flake8 pytest flake8-docstrings
+        python -m pip install flake8 ruff flake8-docstrings
         pip install -r requirements.txt
         pip install -e .
-    - name: Flake8
+    - name: Flake8 & ruff
       run: |
-        python -m flake8
+        python -m ruff check .
         python -m flake8 --select D --ignore E ibllib/qc/camera.py ibllib/qc/task_metrics.py
     - name: Brainbox tests
       run: |
