+++ ruff.toml
@@ -1,4 +1,23 @@
+preview = true
 line-length = 130
+exclude = [
+    "brainbox/examples/",
+    "examples/"
+]
+
+
+[lint]
+# Enable the same rule sets as flake8
+select = ["E", "F", "W", "W391"]
+
+# Ignore the same rules as in .flake8 config
+ignore = [
+    "E266",  # too many leading '#' for block comment
+    "F841",  # local variable assigned but never used
+]
+
+[lint.extend-per-file-ignores]
+"__init__.py" = ["F401"]  # imported but unused
 
 [format]
 quote-style = "single"
